Move digest to wrapper
