+++ src/wrapper/digest.rs
