+++ src/wrapper/mod.rs
@@ -1,1 +1,2 @@
 pub mod bloom_callback;
+pub mod digest;
